streaming/kafka_producer/requirements.txt
psutil==6.1.1
dill==0.3.9
protobuf==4.24.0
numpy==1.26.4
pandas==2.2.3
pyarrow==18.0.0
requests==2.32.3
feast[postgres,redis]==0.47.0
pyspark==3.5.0
kafka-python==2.0.2